Rename EnumMap to ObjectCounter
