+++ DND-Character-Sheets.vcxproj
@@ -121,7 +121,7 @@
     <ClCompile Include="SkillSet.cpp" />
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="dnd_backend\entities\EnumMap.h" />
+    <ClInclude Include="dnd_backend\entities\ObjectCounter.h" />
     <ClInclude Include="dnd_backend\entities\Attribute.h" />
     <ClInclude Include="dnd_backend\entities\AttributeSet.h" />
     <ClInclude Include="dnd_backend\entities\Currency.h" />
