+++ backend_test/test_resources/EnumMapTest.cpp
@@ -1,5 +1,5 @@
 #include "pch.h"
-#include "EnumMap.h"
+#include "ObjectCounter.h"
 #include "Currency.h"
 
 using namespace DND::entity_details;
