+++ dnd_backend/entities/ObjectCounter.h
